feature: console interface

+++ src/Calculator.java
@@ -1,21 +1,21 @@
 public class Calculator {
-    public int add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 
-    public int subtract(int a, int b) {
+    public double subtract(double a, double b) {
         return a - b;
     }
 
-    public int multiply(int a, int b) {
+    public double multiply(double a, double b) {
         return a * b;
     }
 
-    public double divide(int a, int b) {
+    public double divide(double a, double b) {
         if (b == 0) {
             throw new IllegalArgumentException("Division by zero is not allowed");
         }
-        return (double) a / b;
+        return a / b;
     }
 }
